+++ contracts/adapters/interfaces/IManaging.sol
@@ -40,9 +40,9 @@ interface IManaging {
 
     function sponsorProposal(
         DaoRegistry dao,
-        uint64 proposalId,
+        uint256 proposalId,
         bytes calldata data
     ) external;
 
-    function processProposal(DaoRegistry dao, uint64 proposalId) external;
+    function processProposal(DaoRegistry dao, uint256 proposalId) external;
 }
